incrementally "minify" code map, refactor
Code maps are minified based on some token limit. The
minification process happens as the maps are generated in the event that
we don't try to generate mapes if a user forgot to add some large
directory to their .gitignore.

The code map's will be created and minified in the following order:

1. Full map
2. No signatures
3. File-structure only
4. No map