Partial files allowed
The user can specify a partial file with : syntax e.g. test.txt:1-2,5
will put only lines 1,2 and 5 in scope. It is checked if that file with
a colon already exists. For this a File object is created in the
code_file_manager which has the path and segment information bundled.

The expand function is moved from app.py to code_file_manager.py

Test with:
`mentat README.md:6-12,16`
> Replace all instance of the word mentat with test.