Fix: cli flag parsing
The boolean flags would still be set to true or false if no flag was
added so settings set in the config file would be overwritten with no
flag. Namespace values are now not stored if they are equal to the
default.

Also I realized attrs provided a boolean converter so I removed mine.