Benchmark Result Summary Introduced
The goal is for this to be reused by the new benchmark runner. A
benchmark results fields can specify how they are aggregated in which
case their total/count/percent true/histogram will be displayed in the
header. Or not in which case they will be viewable if you click through
to the benchmark.

TODO: actually use it in the benchmark runner. Have indentation/syntax
errors be aggregated.