Benchmark_repos specified with a relative path
The motivation for this PR is a fix for the benchmark runner. Now that
the benchmarks are in a different package the aren't importing mentat
code from where they did previously in the file system.
This change should have no effect for people calling clone_repo from the
mentat root dir.
I think in general it makes sense that the caller of setup_repo would
expect the repo to be places in benchmark_repos of their cwd not
relative to some script that they may not know about and may be in a
strange unpredictable place depending on how mentat is installed.