Benchmark list page
Benchmark Result Summary refactored into Benchmark Run which contains a
list of all the results and a summary. These two objects are stored
seperatly as only the summary is necessary to generate the list. The
summaries are used to generate a list and plotly graph which is synced
to index.html by the runner script. Benchmark specific templates are
moved to the benchmarks module.