Conversation get_messages gets code_message
Frequently in the codebase people will go to conversation for the
messages and then independently get the code_message. It's simpler if
the conversation handles that.

The sampler and reviser both do this but with a different system prompt.
I think we should also change get_conversation to have an optional
system prompt parameter that overrides the default parser system prompt
when passed.

We should also provide a get_token_count method as many callers are
getting the messages just for this one piece of information.